evry time make great job again

+++ data/Inner/AdminInnerAPI.py
@@ -1,7 +1,5 @@
 import datetime
-from flask import jsonify
 from data.user import User
-from data.API.AdminAPI.parser_admin import parser_admin
 from data.API.AuditlogAPI.AuditlogResource import add_auditlog
 from data.Inner.main_file import raise_error, check_admin_status, check_admin
 
@@ -27,35 +25,35 @@ def check_password(password):
     errors = {0: 'Пароль должен быть в длину 8 или более символов', 1: 'Пароль должен содержать хотя бы 1 букву',
               2: 'Пароль должен содержать хотя бы 1 цифру'}
     if not len(password) >= 8:
-        raise_error(errors[0])
+        return raise_error(errors[0])
     if password.isdigit():
-        raise_error(errors[1])
+        return raise_error(errors[1])
     if password.isalpha():
-        raise_error(errors[2])
+        return raise_error(errors[2])
     return True
 
 
 def find_by_id(id, session, status=0):
     user = session.query(User).get(id)
     if not user:
-        raise_error(f"Пользователь не найден", session)
+        return raise_error(f"Пользователь не найден", session), 1
     if user.status >= status or status < 1:
-        raise_error("У вас недостаточно прав для этого", session)
+        return raise_error("У вас недостаточно прав для этого", session), 1
     return user, session
 
 
 def edit_admin(args):
     count, f = 0, False
     if not all(args[key] is not None for key in ['admin_email', 'action']):
-        raise_error("Отсутствуют важные параметры")
+        return raise_error("Отсутствуют важные параметры")
     admin, session = check_admin(args['admin_email'])
     if args["action"] == "get":
         session.close()
-        return jsonify(admin.to_dict(only=('id', 'name', 'surname', 'status', 'email', 'created_date')))
+        return admin.to_dict(only=('id', 'name', 'surname', 'status', 'email', 'created_date'))
     elif args["action"] == "get_list":
         admins = session.query(User).all()
         session.close()
-        return jsonify([item.to_dict(only=('id', 'surname', 'name', 'status', 'email', 'created_date')) for item in admins])
+        return [item.to_dict(only=('id', 'surname', 'name', 'status', 'email', 'created_date')) for item in admins]
     elif args["action"] == "delete":
         pass
     elif args["action"] == "put":
@@ -65,7 +63,7 @@ def edit_admin(args):
             count += 1
             if key == 'email':
                 if session.query(User).filter(User.email == args["email"]).first():
-                    raise_error("Этот email уже занят", session)
+                    return raise_error("Этот email уже занят", session)
                 admin.email = args['email']
             if key == 'name':
                 admin.name = args["name"]
@@ -73,7 +71,7 @@ def edit_admin(args):
                 admin.surname = args["surname"]
         if args["change_password"]:
             if not admin.check_password(args["check_admin_password"]):
-                raise_error("Пароль не совпадает с текущим паролем", session)
+                return raise_error("Пароль не совпадает с текущим паролем", session)
             check_password(args['new_admin_password'])
             admin.set_password(args['new_admin_password'])
             f, count = True, count + 1
@@ -87,26 +85,28 @@ def edit_admin(args):
         add_auditlog("Изменение", f"Пользователь {admin.name} {admin.surname} изменяет сам себя: {', '.join(list_chang)}", admin,
                      datetime.datetime.now())
         session.close()
-        return jsonify({"success": f"Пользователь {admin.name} {admin.surname} успешно изменён"})
-    raise_error("Неизвестный запрос", session)
+        return {"success": f"Пользователь {admin.name} {admin.surname} успешно изменён"}
+    return raise_error("Неизвестный запрос", session)
 
 
 def edit_admin_admin(user_id, args):
     count, f = 0, False
-    if not all(args[key] is not None for key in ['admin_email', 'action', 'admin_password']):  # При изменении себя/админов запрашивать пароль!!!!!!!!!!!
-        raise_error("Отсутствуют важные параметры")
+    if not all(args[key] is not None for key in ['admin_email', 'action']):  # При изменении себя/админов запрашивать пароль!!!!!!!!!!!
+        return raise_error("Отсутствуют важные параметры")
     admin, session = check_admin_status(args['admin_email'], 1)
     user, session = find_by_id(user_id, session, admin.status)
+    if type(user) == dict:
+        return user
     if args["action"] == "get":
         session.close()
-        return jsonify(user.to_dict(only=('id', 'surname', 'name', 'status', 'email', 'created_date')))
+        return user.to_dict(only=('id', 'surname', 'name', 'status', 'email', 'created_date'))
     elif args["action"] == "delete":
         session.delete(user)
         session.commit()
         add_auditlog("Удаление", f"Админ {admin.name} {admin.surname} удаляет админа {user.name} {user.surname}",
                      admin, datetime.datetime.now())
         session.close()
-        return jsonify({"success": f"Пользователь {user.name} {user.surname} успешно удалён"})
+        return {"success": f"Пользователь {user.name} {user.surname} успешно удалён"}
     elif args["action"] == "put":
         user_dict = user.to_dict(only=('id', 'name', 'surname', 'status', 'email'))
         keys = list(filter(lambda key: args[key] is not None and key in user_dict and args[key] != user_dict[key], list(args.keys())))
@@ -115,7 +115,7 @@ def edit_admin_admin(user_id, args):
                 count += 1
                 if key == 'email':
                     if session.query(User).filter(User.id == args["email"]).first():
-                        raise_error("Этот email уже занят", session)
+                        return raise_error("Этот email уже занят", session)
                     user.email = args['email']
                 if key == 'name':
                     user.name = args["name"]
@@ -123,11 +123,11 @@ def edit_admin_admin(user_id, args):
                     user.surname = args["surname"]
                 if key == 'status':
                     if admin.status < args['status']:
-                        raise_error("У вас недостаточно прав для этого", session)
+                        return raise_error("У вас недостаточно прав для этого", session)
                     user.status = args["status"]
-        if args["change_password"]:
+        if "change_password" in args:
             if not admin.check_password(args["check_admin_password"]):
-                raise_error("Пароль не совпадает с текущим паролем", session)
+                return raise_error("Пароль не совпадает с текущим паролем", session)
             check_password(args['new_admin_password'])
             user.set_password(args['new_admin_password'])
             f, count = True, count + 1
@@ -142,17 +142,16 @@ def edit_admin_admin(user_id, args):
                      f"Админ {admin.name} {admin.surname} изменяет пользователя {user.name} {user.surname}: {', '.join(list_chang)}",
                      admin, datetime.datetime.now())
         session.close()
-        return jsonify({"success": f"Пользователь {user.name} {user.surname} успешно изменён"})
-    session.close()
-    raise_error("Неизвестный запрос")
+        return {"success": f"Пользователь {user.name} {user.surname} успешно изменён"}
+    return raise_error("Неизвестный запрос", session)
 
 
 def create_admin(args):
-    if not all(args[key] is not None for key in ['surname', 'name', 'email', 'admin_email', 'admin_password']):
-        raise_error('Пропущены некоторые аргументы, необходимые для создания пользователя')
-    admin, session = check_admin_status(args['admin_email'], args["admin_password"])
+    if not all(args[key] is not None for key in ['surname', 'name', 'email', 'admin_email']):
+        return raise_error('Пропущены некоторые аргументы, необходимые для создания пользователя')
+    admin, session = check_admin_status(args['admin_email'])
     if session.query(User).filter(User.email == args['email']).first():
-        raise_error("Этот email уже занят", session)
+        return raise_error("Этот email уже занят", session)
     check_password(args["new_admin_password"])
     new_admin = User()
     new_admin.name = args["name"]
@@ -163,7 +162,7 @@ def create_admin(args):
         if admin.status > args['status']:
             new_admin.status = args['status']
         else:
-            raise_error("Слишком высокий статус нового админа", session)
+            return raise_error("Слишком высокий статус нового админа", session)
     else:
         new_admin.status = 0
     new_admin.created_date = datetime.datetime.now()
@@ -173,4 +172,4 @@ def create_admin(args):
                  f"Админ {admin.name} {admin.surname} создаёт админа {new_admin.name} {new_admin.surname}: {new_admin.to_dict(only=('id', 'name', 'surname', 'status', 'email'))}",
                  admin, datetime.datetime.now())
     session.close()
-    return jsonify({'success': f'Пользователь {new_admin.name} {new_admin.surname} создан'})
+    return {'success': f'Пользователь {new_admin.name} {new_admin.surname} создан'}
